feat:added notifications feature

+++ src/components/posts/actions.ts
@@ -56,20 +56,65 @@ export async function getPostLikes(postId: string) {
 export async function likePost(postId: string) {
   const { user: loggedInUser } = await validateRequest();
   if (!loggedInUser) throw Error("Unauthenticated");
+
+  const post = await db.post.findUnique({
+    where: { id: postId },
+    select: {
+      userId: true,
+    },
+  });
+
+  if (!post) throw Error("Post not found");
   try {
-    await db.like.upsert({
-      where: {
-        userId_postId: {
+    await db.$transaction([
+      db.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
+          },
+        },
+        create: {
           userId: loggedInUser.id,
           postId,
         },
-      },
-      create: {
-        userId: loggedInUser.id,
-        postId,
-      },
-      update: {},
-    });
+        update: {},
+      }),
+      ...(loggedInUser.id !== post.userId
+        ? [
+            db.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ]);
+    // await db.like.upsert({
+    //   where: {
+    //     userId_postId: {
+    //       userId: loggedInUser.id,
+    //       postId,
+    //     },
+    //   },
+    //   create: {
+    //     userId: loggedInUser.id,
+    //     postId,
+    //   },
+    //   update: {},
+    // });
+
+    // await db.notification.create({
+    //   data: {
+    //     issuerId: loggedInUser.id,
+    //     recipientId: post.userId,
+    //     postId,
+    //     type: "LIKE",
+    //   },
+    // });
     return { message: "success" };
   } catch (error) {
     console.error(error);
@@ -80,12 +125,29 @@ export async function likePost(postId: string) {
 export async function deleteLike(postId: string) {
   const { user: loggedInUser } = await validateRequest();
   if (!loggedInUser) throw Error("Unauthenticated");
+  const post = await db.post.findUnique({
+    where: { id: postId },
+    select: {
+      userId: true,
+    },
+  });
+  if (!post) throw Error("Post not found");
+
   try {
-    await db.like.deleteMany({ where: { userId: loggedInUser.id, postId } });
+    await db.$transaction([
+      db.like.deleteMany({ where: { userId: loggedInUser.id, postId } }),
+      db.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ]);
     return { message: "success" };
   } catch (error) {
     console.error(error);
     throw Error(`${error}`);
   }
-  
 }
